Added in this version: -Mapping one-to-many and many-to-one -Name field to channel table -Jackson annotation on DTO’s to avoid recursivity on json objects -Insert statements on initial_data

+++ app-tdt/src/main/java/com/efe13/tdt/controller/PopulationController.java
@@ -30,7 +30,7 @@ public class PopulationController {
 	}
 	
 	@RequestMapping( value="/{id}", method=RequestMethod.GET )
-	public DTOAPI getPopultation( @PathVariable("id") short populationId ) {
+	public DTOAPI getPopultation( @PathVariable("id") int populationId ) {
 		PopulationDTO population = new PopulationDTO();
 		population.setId( populationId );
